fix: update write support when using Lance namespace (#40)
This PR makes a few changes to the write path, to support having a Lance
namespace provide location and storage options. When writing in `create`
mode, an empty table will be created, the location and storage options
retrieved, and then the data written to the location. If an error occurs
when writing the data, you will end up with an empty table, which isn't
ideal. When transaction support is fleshed out in the namespace API,
this can be moved to a create transaction.

In addition, the namespace was removed as a member variable in the sink,
as that cannot be serialized. Related methods that relied on it were
removed also.

Finally, a small change was made to the source so it can support new
versions of Ray (v2.48+).

This addresses https://github.com/lancedb/lance-ray/issues/38.

---------

Co-authored-by: Bryan Keller <bkeller@netflix.com>