refactor: remove numpy dependency by using itertools.batched/more_itertools.divide (#63)
As far as I can see, the only reason why `lance-ray` depends on `numpy`
is that it uses the `array_split` method from it.

IMHO, just for one function, that's a way too huge dependency to take
in. In our specific case, we're unfortunately still limited to `numpy` <
`2.0`, and that `lance-ray` depends on `numpy` >= `2.0.0` is a problem
for us.

I think `numpy` < `2.0` also contained an `array_split` method, so in
principle one could have lifted the version restriction a bit to fix our
problem, but I refactored this directly a bit to not depend on `numpy`
at all anymore.

---------

Co-authored-by: Jonas Dedden <jonas.dedden@deepl.com>